Separate server.js into initializeApp
